move device code screen to modal window
