+++ packages/tokens-studio-for-figma/src/app/components/Settings/index.tsx
