+++ packages/tokens-studio-for-figma/src/app/components/Subscription/index.tsx
@@ -1,0 +1,3 @@
+import Subscription from './Subscription';
+
+export default Subscription;
